split up node/links in rust
