+++ oceanics-io-api-rust/src/lib.rs
@@ -1,4 +1,7 @@
 mod cypher;
+mod node;
+mod links;
+
 use wasm_bindgen::prelude::*;
 
 extern crate console_error_panic_hook;
